requirements.txt
asgiref==3.7.2
Django==5.0.2
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
psycopg[binary]==3.2.3
PyJWT==2.8.0
pytz==2024.1
sqlparse==0.4.4
django-environ
pillow
setuptools
drf-yasg